refactor: 가독성을 높이기 위해 메서드 분리 및 수정
