--- src/main/java/lotto/converter/Converter.java
@@ -1,7 +1,0 @@
-package lotto.converter;
-
-@FunctionalInterface
-public interface Converter<S, T> {
-
-    T convert(S source);
-}
