--- src/main/java/lotto/converter/LottoRankingToStringConverter.java
@@ -1,15 +1,0 @@
-package lotto.converter;
-
-import lotto.domain.LottoRanking;
-
-import java.util.List;
-
-public class LottoRankingToStringConverter implements Converter<List<LottoRanking>, List<String>> {
-
-    @Override
-    public List<String> convert(List<LottoRanking> lottoRankings) {
-        return lottoRankings.stream()
-                .map(Enum::name)
-                .toList();
-    }
-}
