+++ src/test/java/lotto/service/StatisticsServiceTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
+import java.util.Map;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -25,9 +26,9 @@ class StatisticsServiceTest {
     @Test
     @DisplayName("사용자가 구매한 총 로또와 당첨 번호를 비교하여 전체 당첨 내역을 알 수 있다.")
     void checkLottoResult() {
-        LottoResult lottoResult = statisticsService.checkLottoResult(winningLotto, userLottos);
+        Map<LottoRanking, Integer> result = statisticsService.checkLottoResult(winningLotto, userLottos);
 
-        assertThat(lottoResult.getResult()).contains(
+        assertThat(result).contains(
                 entry(LottoRanking.FIFTH, 1),
                 entry(LottoRanking.NOTHING, 7)
         );
@@ -36,7 +37,7 @@ class StatisticsServiceTest {
     @Test
     @DisplayName("사용자가 구매한 로또 번호의 결과를 통해 수익률을 계산할 수 있다.")
     void calculateRateOfReturn() {
-        LottoResult lottoResult = statisticsService.checkLottoResult(winningLotto, userLottos);
+        Map<LottoRanking, Integer> lottoResult = statisticsService.checkLottoResult(winningLotto, userLottos);
         double result = statisticsService.calculateRateOfReturn(lottoResult, userLottos);
 
         assertThat(result).isEqualTo(62.5);
